SWM-92 refactor: Entity Id 타입 String으로 변경
