+++ src/main/java/swm/betterlife/antifragile/domain/auth/dto/request/LogoutRequest.java
@@ -1,4 +1,4 @@
-package swm.betterlife.antifragile.domain.auth.dto;
+package swm.betterlife.antifragile.domain.auth.dto.request;
 
 public record LogoutRequest(
     String refreshToken
