+++ src/main/java/swm/betterlife/antifragile/domain/auth/dto/response/LoginResponse.java
@@ -1,9 +1,9 @@
-package swm.betterlife.antifragile.domain.auth.dto;
+package swm.betterlife.antifragile.domain.auth.dto.response;
 
 import lombok.Builder;
-import swm.betterlife.antifragile.domain.member.dto.MemberDetailResponse;
+import swm.betterlife.antifragile.domain.member.dto.response.MemberDetailResponse;
 import swm.betterlife.antifragile.domain.member.entity.Member;
-import swm.betterlife.antifragile.domain.token.dto.TokenIssueResponse;
+import swm.betterlife.antifragile.domain.token.dto.response.TokenIssueResponse;
 
 @Builder
 public record LoginResponse(
