+++ src/main/java/swm/betterlife/antifragile/domain/auth/dto/response/SignUpResponse.java
@@ -1,4 +1,4 @@
-package swm.betterlife.antifragile.domain.auth.dto;
+package swm.betterlife.antifragile.domain.auth.dto.response;
 
 import swm.betterlife.antifragile.domain.member.entity.LoginType;
