+++ src/main/java/swm/betterlife/antifragile/domain/auth/service/AuthService.java
@@ -12,15 +12,15 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import swm.betterlife.antifragile.common.jwt.util.JwtProvider;
-import swm.betterlife.antifragile.domain.auth.dto.LoginRequest;
-import swm.betterlife.antifragile.domain.auth.dto.LoginResponse;
-import swm.betterlife.antifragile.domain.auth.dto.LogoutRequest;
-import swm.betterlife.antifragile.domain.auth.dto.SignUpRequest;
-import swm.betterlife.antifragile.domain.member.dto.MemberDetailResponse;
+import swm.betterlife.antifragile.domain.auth.dto.request.LoginRequest;
+import swm.betterlife.antifragile.domain.auth.dto.request.LogoutRequest;
+import swm.betterlife.antifragile.domain.auth.dto.request.SignUpRequest;
+import swm.betterlife.antifragile.domain.auth.dto.response.LoginResponse;
+import swm.betterlife.antifragile.domain.member.dto.response.MemberDetailResponse;
 import swm.betterlife.antifragile.domain.member.entity.LoginType;
 import swm.betterlife.antifragile.domain.member.entity.Member;
 import swm.betterlife.antifragile.domain.member.repository.MemberRepository;
-import swm.betterlife.antifragile.domain.token.dto.TokenIssueResponse;
+import swm.betterlife.antifragile.domain.token.dto.response.TokenIssueResponse;
 import swm.betterlife.antifragile.domain.token.service.TokenService;
 
 @Slf4j
