+++ src/main/java/swm/betterlife/antifragile/domain/content/entity/Content.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import swm.betterlife.antifragile.common.entity.BaseTimeEntity;
@@ -19,11 +18,11 @@ import swm.betterlife.antifragile.common.entity.BaseTimeEntity;
 public class Content extends BaseTimeEntity {
 
     @Id
-    private ObjectId id;
+    private String id;
 
     private String title;
 
-    private List<ObjectId> likeMemberIds = new ArrayList<>();
+    private List<String> likeMemberIds = new ArrayList<>();
 
     private List<SaveMember> saveMembers = new ArrayList<>();
