+++ src/main/java/swm/betterlife/antifragile/domain/content/entity/SaveMember.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.bson.types.ObjectId;
 
 @Getter
 @NoArgsConstructor
@@ -14,7 +13,7 @@ import org.bson.types.ObjectId;
 @Builder
 public class SaveMember {
 
-    private ObjectId memberId;
+    private String memberId;
 
     private List<String> tags = new ArrayList<>();
 }
