+++ src/main/java/swm/betterlife/antifragile/domain/diaryanalysis/entity/DiaryAnalysis.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import swm.betterlife.antifragile.common.entity.BaseTimeEntity;
@@ -19,9 +18,9 @@ import swm.betterlife.antifragile.common.entity.BaseTimeEntity;
 public class DiaryAnalysis extends BaseTimeEntity {
 
     @Id
-    private ObjectId id;
+    private String id;
 
-    private ObjectId memberId;
+    private String memberId;
 
     private List<String> emotions = new ArrayList<>();
