+++ src/main/java/swm/betterlife/antifragile/domain/diaryanalysis/entity/RecommendContent.java
@@ -1,11 +1,9 @@
 package swm.betterlife.antifragile.domain.diaryanalysis.entity;
 
-import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.bson.types.ObjectId;
 
 @Getter
 @NoArgsConstructor
@@ -13,7 +11,7 @@ import org.bson.types.ObjectId;
 @Builder
 public class RecommendContent {
 
-    private ObjectId id;
+    private String id;
 
     private String title;
