+++ src/main/java/swm/betterlife/antifragile/domain/diaryanalysis/entity/SelectedEmoticon.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.bson.types.ObjectId;
 
 @Getter
 @NoArgsConstructor
@@ -12,7 +11,7 @@ import org.bson.types.ObjectId;
 @Builder
 public class SelectedEmoticon {
 
-    private ObjectId emoticonThemeId;
+    private String emoticonThemeId;
 
     private String emotion;
 }
