+++ src/main/java/swm/betterlife/antifragile/domain/member/dto/request/NicknameModifyRequest.java
@@ -1,4 +1,4 @@
-package swm.betterlife.antifragile.domain.member.dto;
+package swm.betterlife.antifragile.domain.member.dto.request;
 
 public record NicknameModifyRequest(
     String nickname
