+++ src/main/java/swm/betterlife/antifragile/domain/member/dto/request/ProfileImgModifyRequest.java
@@ -1,4 +1,4 @@
-package swm.betterlife.antifragile.domain.member.dto;
+package swm.betterlife.antifragile.domain.member.dto.request;
 
 public record ProfileImgModifyRequest(
     String profileImg
