+++ src/main/java/swm/betterlife/antifragile/domain/member/service/MemberService.java
@@ -3,9 +3,9 @@ package swm.betterlife.antifragile.domain.member.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import swm.betterlife.antifragile.domain.member.dto.MemberDetailResponse;
-import swm.betterlife.antifragile.domain.member.dto.NicknameModifyRequest;
-import swm.betterlife.antifragile.domain.member.dto.ProfileImgModifyRequest;
+import swm.betterlife.antifragile.domain.member.dto.request.NicknameModifyRequest;
+import swm.betterlife.antifragile.domain.member.dto.request.ProfileImgModifyRequest;
+import swm.betterlife.antifragile.domain.member.dto.response.MemberDetailResponse;
 import swm.betterlife.antifragile.domain.member.entity.LoginType;
 import swm.betterlife.antifragile.domain.member.entity.Member;
 import swm.betterlife.antifragile.domain.member.repository.MemberRepository;
