+++ src/main/java/swm/betterlife/antifragile/domain/pointreceipt/entity/PointReceipt.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import swm.betterlife.antifragile.common.entity.BaseTimeEntity;
@@ -17,13 +16,13 @@ import swm.betterlife.antifragile.common.entity.BaseTimeEntity;
 public class PointReceipt extends BaseTimeEntity {
 
     @Id
-    private ObjectId id;
+    private String id;
 
-    private ObjectId memberId;
+    private String memberId;
 
     private PointReceiptType type;
 
     private Integer amount;
 
-    private ObjectId emoticonThemeId;
+    private String emoticonThemeId;
 }
