+++ src/main/java/swm/betterlife/antifragile/domain/token/dto/request/TokenReIssueRequest.java
@@ -1,4 +1,4 @@
-package swm.betterlife.antifragile.domain.token.dto;
+package swm.betterlife.antifragile.domain.token.dto.request;
 
 public record TokenReIssueRequest(
         String refreshToken
