+++ src/main/java/swm/betterlife/antifragile/domain/token/dto/response/TokenIssueResponse.java
@@ -1,4 +1,4 @@
-package swm.betterlife.antifragile.domain.token.dto;
+package swm.betterlife.antifragile.domain.token.dto.response;
 
 public record TokenIssueResponse(
         String accessToken,
